+++ src/core/resources/game_time.rs
@@ -3,10 +3,14 @@ use bevy::{prelude::*, time::Stopwatch};
 #[derive(Resource, Deref)]
 pub struct GameTime(pub Stopwatch);
 
+#[derive(Resource, Deref)]
+pub struct PlayTime(pub Stopwatch);
+
 pub struct GameTimePlugin;
 impl Plugin for GameTimePlugin {
     fn build(&self, app: &mut App) {
         app.insert_resource(GameTime(Stopwatch::new()));
+        app.insert_resource(PlayTime(Stopwatch::new()));
     }
 }
 
@@ -20,4 +24,8 @@ pub(crate) fn tick_game_time(time: Res<Time>, mut game_time: ResMut<GameTime>) {
     // to be. Planets are noticably moving, but not so fast that it *feels*
     // too unnatural. It feels suitable for the game as an arcade shooter,
     // so this is probably where we'll land for v0.1.
+}
+
+pub(crate) fn tick_play_time(time: Res<Time>, mut play_time: ResMut<PlayTime>) {
+    play_time.0.tick(time.delta());
 }